moved SignUp and Login components to views
